Update cRandom.h to specify explicitly the order in which the casting should occur. Improper casting order was causing organism replication to fail (among probably many other things).

+++ source/tools/cRandom.h
@@ -126,7 +126,7 @@ public:
    * @return The pseudo random number.
    * @param max The upper bound for the random numbers (will never be returned).
    **/
-  unsigned int GetUInt(const unsigned int max) { return static_cast<int>(GetDouble()) * max; }
+  unsigned int GetUInt(const unsigned int max) { return static_cast<int>(GetDouble() * static_cast<double>(max)); }
   
   /**
    * Generate an unsigned int out of an interval.
